Extract colorizing shell methods in a new module.
